+++ experimental/psi/psi21/el_q_psi/el_q_psi_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "psi/psi/psi21_experiment/el_mp_psi/el_mp_psi.h"
+#include "experimental/psi/psi21/el_q_psi/el_q_psi.h"
 
 #include <random>
 #include <set>
@@ -26,13 +26,14 @@ namespace psi::psi {
 
 namespace {
 
-struct NMpTestParams {
+struct NCTestParams {
   std::vector<size_t> item_size;
   size_t intersection_size;
+  size_t n;
 };
 
 std::vector<std::vector<std::string>> CreateNPartyItems(
-    const NMpTestParams& params) {
+    const NCTestParams& params) {
   std::vector<std::vector<std::string>> ret(params.item_size.size() + 1);
   ret[params.item_size.size()] =
       test::CreateRangeItems(1, params.intersection_size);
@@ -55,41 +56,65 @@ std::vector<std::vector<std::string>> CreateNPartyItems(
     for (const auto& iter : idx_set) {
       ret[idx][iter] = ret[params.item_size.size()][j++];
     }
+
+    if (/*idx > dis(gen) &&*/ idx >= params.n) {
+      break;
+    }
   }
   return ret;
 }
 
 }  // namespace
 
-class NMpPsiTest : public testing::TestWithParam<NMpTestParams> {};
+class NCPsiTest : public testing::TestWithParam<NCTestParams> {};
 
 // FIXME : this test is not stable in arm env
-TEST_P(NMpPsiTest, Works) {
+TEST_P(NCPsiTest, Works) {
   std::vector<std::vector<std::string>> items;
   std::vector<std::vector<std::string>> resultvec;
   std::vector<std::string> finalresult;
 
   auto params = GetParam();
+  size_t n = params.n;
   items = CreateNPartyItems(params);
   size_t leader_rank = 0;
   uint128_t maxlength = 0;
 
+  if (n >= params.item_size.size()) {
+    SPDLOG_INFO("param error: n > items[0].size() ");
+    return;
+  }
+
+  if (n <= 0) {
+    SPDLOG_INFO("param error: n <= 0 ");
+    return;
+  }
+
+  /*
+  for (size_t j{}; j != items[0].size(); ++j) {
+    SPDLOG_INFO(" items[{}][{}] = {}, size{}", 0, j, items[0][j],
+                items[0].size());
+  }
+  */
   for (size_t i = 0; i < params.item_size.size() - 1; i++) {
     std::vector<std::vector<std::string>> items1;
     items1.push_back(items[0]);
     items1.push_back(items[i + 1]);
     leader_rank = 0;
 
+    /*
+    for (size_t j{}; j != items[i + 1].size(); ++j) {
+         SPDLOG_INFO(" items[{}][{}] = {}, size{}", i+1, j, items[i+1][j],
+         items[i+1].size());
+    }
+    */
+
     auto ctxs = yacl::link::test::SetupWorld(2);
     auto proc = [&](int idx) -> std::vector<std::string> {
-      NmpParty::Options opts;
+      NcParty::Options opts;
       opts.link_ctx = ctxs[idx];
       opts.leader_rank = leader_rank;
-      NmpParty op(opts);
-      // for (size_t j{}; j != items1[idx].size(); ++j) {
-      //   SPDLOG_INFO(" items[{}][{}] = {}, size{}", idx, i, items[idx][i],
-      //   items[idx].size());
-      // }
+      NcParty op(opts);
 
       return op.Run(items[idx]);
     };
@@ -105,26 +130,48 @@ TEST_P(NMpPsiTest, Works) {
     result = f_links[0].get();
     resultvec.push_back(result);
 
-    // for (size_t j = 0; j < result.size() ; j++) {
-    //   SPDLOG_INFO("i{}  j{}, result[j] {}  size{}",i,j,result[j],
-    //   result.size());
-    // }
+    /*
+    for (size_t j = 0; j < result.size(); j++) {
+        SPDLOG_INFO("i{}  j{}, result[j] {}  size{}", i, j, result[j],
+                    result.size());
+    }*/
   }
 
   maxlength = items[0].size();
+  std::vector<uint128_t> qpsivector;
   for (size_t j = 0; j < maxlength; j++) {
+    uint128_t sum = 0;
     for (size_t i = 0; i < params.item_size.size() - 1; i++) {
+      // 如果有的集合没有那么多项就continue
+      // results[i] = f_links[i].get();
       if (resultvec[i].size() <= j) {
-        break;
-      }
-      if (resultvec[i][j] != "0") {
-        break;
+        continue;
       }
 
-      if (i == params.item_size.size() - 2) {
-        finalresult.push_back(items[0][j]);
+      // SPDLOG_INFO(" result[{}][{}] = {}", i, j, resultvec[i][j]);
+      auto it = resultvec[i].begin() + j;
+      std::string element = *it;
+      if (element == "1") {
+        sum++;
       }
     }
+    if (sum >= n) {
+      // todo//推入对应input元素 之后再查输入变量从param中怎么取出推入
+      qpsivector.push_back(1);
+    } else {
+      qpsivector.push_back(0);
+    }
+  }
+
+  // std::vector<std::string> intersectionnparty;
+  for (size_t k{}; k != items[0].size(); ++k) {
+    if (qpsivector[k] == 1) {
+      finalresult.push_back(items[0][k]);
+    }
+  }
+
+  for (size_t i{}; i != finalresult.size(); ++i) {
+    SPDLOG_INFO("intersectionnparty = {}", finalresult[i]);
   }
 
   std::vector<std::string> intersection = items[params.item_size.size()];
@@ -136,14 +183,14 @@ TEST_P(NMpPsiTest, Works) {
 }
 
 INSTANTIATE_TEST_SUITE_P(
-    Works_Instances, NMpPsiTest,
-    testing::Values(NMpTestParams{{0, 3}, 0},                 //
-                    NMpTestParams{{3, 0}, 0},                 //
-                    NMpTestParams{{0, 0}, 0},                 //
-                    NMpTestParams{{4, 3}, 2},                 //
-                    NMpTestParams{{20, 17, 14}, 10},          //
-                    NMpTestParams{{20, 17, 14, 30}, 10},      //
-                    NMpTestParams{{20, 17, 14, 30, 35}, 11},  //
-                    NMpTestParams{{20, 17, 14, 30, 35}, 0}));
-
+    Works_Instances, NCPsiTest,
+    testing::Values(NCTestParams{{0, 3}, 0, 1},                 //
+                    NCTestParams{{3, 0}, 0, 1},                 //
+                    NCTestParams{{0, 0}, 0, 1},                 //
+                    NCTestParams{{4, 3}, 2, 1},                 //
+                    NCTestParams{{20, 17, 14}, 10, 2},          //
+                    NCTestParams{{20, 17, 14, 30}, 10, 3},      //
+                    NCTestParams{{20, 17, 14, 30, 35}, 11, 4},  //
+                    NCTestParams{{20, 17, 14, 30, 35}, 0, 4}));
+// testing::Values(NCTestParams{{3, 0}, 0, 1}));
 }  // namespace psi::psi
